+++ clients/admin-ui/src/features/dataset/DatasetCollectionView.tsx
@@ -3,9 +3,17 @@ import { useRouter } from "next/router";
 import { ChangeEvent, useEffect, useState } from "react";
 
 import { FidesKey } from "../common/fides-types";
+import ColumnDropdown from "./ColumnDropdown";
 import { useGetDatasetByKeyQuery } from "./dataset.slice";
 import DatasetFieldsTable from "./DatasetFieldsTable";
-import { DatasetCollection } from "./types";
+import { ColumnMetadata, DatasetCollection } from "./types";
+
+const ALL_COLUMNS: ColumnMetadata[] = [
+  { name: "Field Name", attribute: "name" },
+  { name: "Description", attribute: "description" },
+  { name: "Personal Data Categories", attribute: "data_categories" },
+  { name: "Identifiability", attribute: "data_qualifier" },
+];
 
 const SuccessMessage = () => (
   <Text>
@@ -31,6 +39,8 @@ const DatasetCollectionView = ({ fidesKey }: Props) => {
   const [activeCollection, setActiveCollection] = useState<
     DatasetCollection | undefined
   >();
+  const [columns, setColumns] = useState<ColumnMetadata[]>(ALL_COLUMNS);
+
   const router = useRouter();
   const toast = useToast();
   const { fromLoad } = router.query;
@@ -75,7 +85,7 @@ const DatasetCollectionView = ({ fidesKey }: Props) => {
 
   return (
     <Box>
-      <Box mb={4} display="flex">
+      <Box mb={4} display="flex" justifyContent="space-between">
         <Select onChange={handleChangeCollection} mr={2} width="auto">
           {collections.map((collection) => (
             <option key={collection.name} value={collection.name}>
@@ -83,9 +93,17 @@ const DatasetCollectionView = ({ fidesKey }: Props) => {
             </option>
           ))}
         </Select>
+        <ColumnDropdown
+          allColumns={ALL_COLUMNS}
+          selectedColumns={columns}
+          onChange={setColumns}
+        />
       </Box>
       {activeCollection && (
-        <DatasetFieldsTable fields={activeCollection.fields} />
+        <DatasetFieldsTable
+          fields={activeCollection.fields}
+          columns={columns}
+        />
       )}
     </Box>
   );