+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The types of changes are:
   * Systems
 * Initial dataset collection UI view
   * Add interaction for viewing a dataset collection
+  * Add column picker
 * Okta, aws and database credentials can now come from `fidesctl.toml` config [#694](https://github.com/ethyca/fides/pull/694)
 
 ### Changed